Co invest (#22)
* Added commitment_type for Pool/Co Invest

* Coinvest first cut

* Added capital commitment to portfolio investments for co invest

* PAsses tests

* PAsses tests

* PAsses tests

* Fixed tests for co invest

* Added hide/show coinvest folio in form

* Added feeder_fund for commitments

* Added type and folio to import of PI for coinvest

* Added better explanation of portfolio attribution

* Added better computation for FIFO

* Backup

* Added better computation for API costs

* Added changes to CD to support co invest

* Added changes to CD to support co invest

* Backup

* Added fund formuals with commitment_type

* Added commitment_type to account entry

* Added commitment_type to account entry

* Added fixes for allocation engine

* Added % for certain account entry

* Added pending call_amount in commitment grid

* Fixed bug with download/print not working for investors documents

* Increased file upload limit to 1Gb

* Round percentage displayed to 2 decimals

* Allow delete of aggregate portfolio

* Allow delete of aggregate portfolio

* Delete all generated account entries

* Fixed bug with import of PI, update to cost_of_sold

* Added Mark Payments Completed to CApitalDistribution

* Added additional checks for co invest portfolio investments

* Added additional checks for co invest portfolio investments

* Added FIFO costs to portfolio view

* Added kyc documents upload

* Added upload of zip documents

* Fixed upload % for import documents

* Added explanation to account entry XL

* Added explanation to account entry XL

* Added validation to prevent CRUD for the formulas

* Added FolderDefaultsJob to cascade folder defaults

* Added FolderDefaultsJob to cascade folder defaults

* Added fund_units to Call and Distribution

* Added better stylesheet and layout to admin screen

* Added import docs to enable zip file orignal

+++ app/packs/misc/import_uploads/jobs/import_unzip_file.rb
@@ -7,7 +7,7 @@ class ImportUnzipFile
       # Unzip the contents into the unzip_dir
       unzip(context.import_file, context.unzip_dir, context.import_upload)
       # We replace the import_file with the index.xlsx, so import pre_process can work right
-      context.import_file = File.open("#{context.unzip_dir}/index.xlsx", "r")
+      context.import_file = File.open("#{context.unzip_dir}/index.xlsx", "r") if File.exist?("#{context.unzip_dir}/index.xlsx")
     else
       context.fail!(message: "Required inputs not present")
     end
